muutokset

+++ src/controllers/user-controller.mjs
@@ -1,5 +1,4 @@
 import bcrypt from 'bcryptjs';
-import { validationResult } from 'express-validator';
 import {
   deleteUserById,
   insertUser,
@@ -7,64 +6,81 @@ import {
   selectUserById,
   updateUserById,
 } from '../models/user-model.mjs';
+import {customError} from '../middlewares/error-handler.mjs';
 
-const getUsers = async (req, res) => {
+/**
+ * Get all users
+ * @param {Object} req - Request object
+ * @param {Object} res - Response object
+ * @param {function} next - next function
+ */
+const getUsers = async (req, res, next) => {
   const result = await listAllUsers();
   if (result.error) {
-    return res.status(result.error).json(result);
+    return next(customError(result, result.error));
   }
   return res.json(result);
 };
 
-const getUserById = async (req, res) => {
+const getUserById = async (req, res, next) => {
   const result = await selectUserById(req.params.id);
   if (result.error) {
-    return res.status(result.error).json(result);
+    return next(customError(result, result.error));
   }
   return res.json(result);
 };
 
-const postUser = async (req, res) => {
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) {
-    return res.status(400).json({ errors: errors.array() });
-  }
-
-  const { username, password, email } = req.body;
+const postUser = async (req, res, next) => {
+  const {username, password, email} = req.body;
   const salt = await bcrypt.genSalt(10);
   const hashedPassword = await bcrypt.hash(password, salt);
-  const result = await insertUser({ username, email, password: hashedPassword });
-
-  if (result.error) {
-    return res.status(500).json(result);
-  }
+  const result = await insertUser(
+    {
+      username,
+      email,
+      password: hashedPassword,
+    },
+    next,
+  );
   return res.status(201).json(result);
 };
 
-const putUser = async (req, res) => {
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) {
-    return res.status(400).json({ errors: errors.array() });
-  }
-
-  const { username, password, email } = req.body;
+const putUser = async (req, res, next) => {
+  // Get userinfo from req.user object extracted from token
+  // Only user authenticated by token can update own data
+  // TODO: admin user can update any user (incl. user_level)
   const userId = req.user.user_id;
+  const {username, password, email} = req.body;
+  // hash password if included in request
   const salt = await bcrypt.genSalt(10);
   const hashedPassword = await bcrypt.hash(password, salt);
-  const result = await updateUserById({ userId, username, password: hashedPassword, email });
-
+  const result = await updateUserById({
+    userId,
+    username,
+    password: hashedPassword,
+    email,
+  });
   if (result.error) {
-    return res.status(500).json(result);
+    return next(customError(result, result.error));
   }
-  return res.status(201).json(result);
+  return res.status(200).json(result);
 };
 
-const deleteUser = async (req, res) => {
+const deleteUser = async (req, res, next) => {
+  // console.log('deleteUser', req.user, req.params.id);
+  // admin user can delete any user
+  // user authenticated by token can delete itself
+  if (
+    req.user.user_level !== 'admin' &&
+    req.user.user_id !== parseInt(req.params.id)
+  ) {
+    return next(customError('Unauthorized', 401));
+  }
   const result = await deleteUserById(req.params.id);
   if (result.error) {
-    return res.status(500).json(result);
+    return next(customError(result, result.error));
   }
   return res.json(result);
 };
 
-export { getUsers, getUserById, postUser, putUser, deleteUser };
+export {getUsers, getUserById, postUser, putUser, deleteUser};
